+++ plugins/dotplot-view/src/DotplotView/components/ImportForm/index.tsx
@@ -20,7 +20,7 @@ import { getSession, isSessionWithAddTracks } from '@jbrowse/core/util'
 import { ErrorMessage, AssemblySelector } from '@jbrowse/core/ui'
 
 // locals
-import { DotplotViewModel } from '../model'
+import { DotplotViewModel } from '../../model'
 import ImportCustomTrack from './ImportCustomTrack'
 import ImportSyntenyTrackSelector from './ImportSyntenyTrackSelector'
