+++ plugins/dotplot-view/src/DotplotView/components/util.ts
@@ -111,27 +111,3 @@ export function makeTicks(
   }
   return ticks
 }
-
-export function getName(
-  sessionTrackData?: { uri: string } | { localPath: string } | { name: string },
-) {
-  return sessionTrackData
-    ? // @ts-expect-error
-      sessionTrackData.uri ||
-        // @ts-expect-error
-        sessionTrackData.localPath ||
-        // @ts-expect-error
-        sessionTrackData.name
-    : undefined
-}
-
-export function stripGz(fileName: string) {
-  return fileName.endsWith('.gz') ? fileName.slice(0, -3) : fileName
-}
-
-export function basename(str: string) {
-  return str.slice(0, str.indexOf('str'))
-}
-export function extName(str: string) {
-  return str.split('.').slice(0, -1).join('.')
-}
